backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-yaml==2.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
